fixes and readability improvement
- add filters for `Components` and `Systems`
- replace `float` in `Position` and `Size` to int as SDL works with int
- use `const ... &` instead of `... const&` in Game.h
- move `GameObject` dtor to the ctor
- rename `dt` to `deltaTime` in `RenderSystem`
- add blank lines between `pragma` and `include` and between `include` and  code

+++ BloomFramework/include/Components/Position.h
@@ -4,8 +4,8 @@
 
 namespace bloom {
 	struct Position {
-		Position(float xPos = 0.0f, float yPos = 0.0f) : x(xPos), y(yPos) {}
+		Position(int xPos = 0, int yPos = 0) : x(xPos), y(yPos) {}
 
-		float x, y;
+		int x, y;
 	};
 }
