+++ BloomFramework/include/Components/Size.h
@@ -4,8 +4,8 @@
 
 namespace bloom {
 	struct Size {
-		Size(float w = 1.0f, float h = 1.0f) : w(w), h(h) {}
+		Size(int w = 1, int h = 1) : w(w), h(h) {}
 
-		float w, h;
+		int w, h;
 	};
 }
