+++ Test Bench/GameObjectTest/TestGameObject.h
@@ -8,8 +8,8 @@ public:
 	using bloom::GameObject::GameObject;
 
 	void init(SDL_Rect pos_and_size = SDL_Rect{ 50,50, 256, 256 }, const std::string & texturePath = "Assets/TestChar.png", std::optional<SDL_Rect> srcRect = std::nullopt) {
-		m_registry.replace<bloom::Position>(m_entity, static_cast<float>(pos_and_size.x), static_cast<float>(pos_and_size.y));
-		m_registry.assign<bloom::Size>(m_entity, static_cast<float>(pos_and_size.w), static_cast<float>(pos_and_size.h));
+		m_registry.replace<bloom::Position>(m_entity, pos_and_size.x, pos_and_size.y);
+		m_registry.assign<bloom::Size>(m_entity, pos_and_size.w, pos_and_size.h);
 		auto tmp = m_gameInstance->textures.load(texturePath);
 
 		m_registry.assign<bloom::Sprite>(m_entity, tmp, srcRect);
